refactor(aether-cli): Move a bunch of files into another module dir
