+++ crates/aether-cli/src/acp/session/mod.rs
@@ -1,0 +1,13 @@
+pub(crate) mod actor;
+pub(crate) mod agent_key;
+pub(crate) mod agents;
+pub(crate) mod config;
+pub(crate) mod config_setting;
+pub(crate) mod error;
+pub(crate) mod factory;
+pub(crate) mod model;
+pub(crate) mod registry;
+pub(crate) mod runtime;
+pub(crate) mod slash_commands;
+
+pub(crate) use registry::SessionRegistry;
