+++ crates/aether-cli/src/acp/testing.rs
@@ -1,12 +1,12 @@
 use super::agent::acp_agent_builder;
-use super::agent_key::AgentKey;
-use super::agent_runtime::{AgentRuntime, RuntimeEvent, RuntimeFactory};
-use super::error::SessionError;
 use super::fake_prompt_mcp::FakePromptMcp;
-use super::model_config::{Modes, ValidatedMode};
-use super::session_actor::{SessionActor, SessionActorInit};
-use super::session_agents::SessionAgents;
-use super::session_config_state::SessionConfigState;
+use super::session::actor::{SessionActor, SessionActorInit};
+use super::session::agent_key::AgentKey;
+use super::session::agents::SessionAgents;
+use super::session::config::SessionConfigState;
+use super::session::error::SessionError;
+use super::session::model::{Modes, ValidatedMode};
+use super::session::runtime::{AgentRuntime, RuntimeEvent, RuntimeFactory};
 use super::state::{AcpState, AcpStateConfig};
 use crate::error::CliError;
 use crate::resolve::InitialSessionSelection;
